Moving class InMemoryFeedStore to production
